requirements.txt
openai==0.28.0
streamlit==1.42.0
requests==2.28.1
python-dotenv==1.0.0